Trying to add Google Maps to front page
And failing. It probably has to do with the way I configured the addons/plugins
for broccoli-sass and ember-cli-sass. Regardless the exact problem I'm having
is
C:\Users\Colby\Documents\GitHub\meteorite-map\tmp\broccoli_merge_trees-output_path-ldNhIf2I.tmp\app\styles\app.[scss|sass] does not exist
So I've tried modifying the output paths in ember-cli-build.js to no avail.
ember-cli-sass relies on broccoli-sass, so I can't figure out if the
problem lies with the ember plugin or the broccoli plugin. I'm almost about
ready to just let the SASS gem on my computer take care of the problem
for me and just have it output the css file to a spot that ember likes.