Started adding Google Maps as component
Steps to recreate are as follows:
Run `ember generate component google-maps` in the console
Mostly follow the instructions at:
http://blog.atmartin.io/google-maps-integration-ember-cli-edition/

This should work except for the fact that `maps` is undefined in
`window.google.maps.LatLng()` and probably
`window.google.maps.Map()`. From the looks of it in the error console,
JQuery is the one firing off the error, though I'm probably mistaken.